:truck: Rename : Main 컴포넌트를 page 로 이동
- Main.tsx 를 Mainpage.tsx 로 변경 후 page 폴더로 이동
- Main 컴포넌트를 PostList 컴포넌트로 변경

Issuse : #8
