+++ client/src/page/Mainpage.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 import { styled } from "styled-components";
 
-import { mainData } from "./DummyData";
-import Post from "./Post";
-import AskQuestionButton from "../AskQuestionButton/Index";
-import LeftSidebar from "../LeftSidebar";
-import RightSidebar from "../RightSidebar/Index";
+import { mainData } from "../components/PostList/DummyData";
+import Post from "../components/PostList/Index";
+import AskQuestionButton from "../components/AskQuestionButton/Index";
+import LeftSidebar from "../components/LeftSidebar";
+import RightSidebar from "../components/RightSidebar/Index";
 
 const title = "Top Questions";
 const filter = ["Interesting", "Bountied", "Hot", "Week", "Month"];
 
-const Main = () => {
+const Mainpage = () => {
   return (
     <>
       <MainContainer>
@@ -43,7 +43,7 @@ const Main = () => {
   );
 };
 
-export default Main;
+export default Mainpage;
 
 const MainContainer = styled.div`
   width: 80%;
